+++ src/main/java/duke/Task.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke;
 
 
 public class Task {
